requirement.txt
requests
psycopg2-binary 
trino
pydantic
pandas
matplotlib
seaborn
psycopg2-binary